+++ nym-authenticator-client/src/lib.rs
@@ -1,16 +1,13 @@
 // Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
+use nym_authenticator_requests::client_message::QueryMessageImpl;
 use nym_bandwidth_controller::{BandwidthTicketProvider, DEFAULT_TICKETS_TO_SPEND};
-use nym_registration_common::{
-    GatewayData, DEFAULT_PRIVATE_ENTRY_WIREGUARD_KEY_FILENAME,
-    DEFAULT_PRIVATE_EXIT_WIREGUARD_KEY_FILENAME, DEFAULT_PUBLIC_ENTRY_WIREGUARD_KEY_FILENAME,
-    DEFAULT_PUBLIC_EXIT_WIREGUARD_KEY_FILENAME,
-};
-use rand::rngs::OsRng;
+use nym_crypto::asymmetric::x25519::KeyPair;
+use nym_registration_common::GatewayData;
 use std::net::{IpAddr, SocketAddr};
-use std::path::PathBuf;
 use std::str::FromStr;
+use std::sync::Arc;
 use std::time::Duration;
 use tracing::{debug, error, trace};
 
@@ -19,20 +16,16 @@ use nym_authenticator_requests::{
     client_message::ClientMessage, response::AuthenticatorResponse, traits::Id, v2, v3, v4, v5,
     AuthenticatorVersion,
 };
-use nym_credentials_interface::TicketType;
-use nym_crypto::asymmetric::x25519::KeyPair;
-use nym_pemstore::KeyPairPath;
+use nym_credentials_interface::{CredentialSpendingData, TicketType};
 use nym_sdk::mixnet::{IncludedSurbs, Recipient};
 use nym_service_provider_requests_common::{Protocol, ServiceProviderTypeExt};
 use nym_wireguard_types::PeerPublicKey;
 
 mod error;
 mod helpers;
-mod legacy;
 mod mixnet_listener;
 
 pub use crate::error::{Error, Result};
-pub use crate::legacy::LegacyAuthenticatorClient;
 pub use crate::mixnet_listener::{AuthClientMixnetListener, AuthClientMixnetListenerHandle};
 
 pub struct AuthenticatorClient {
@@ -42,34 +35,21 @@ pub struct AuthenticatorClient {
     pub auth_recipient: Recipient,
     auth_version: AuthenticatorVersion,
 
-    keypair: KeyPair,
+    keypair: Arc<KeyPair>,
     ip_addr: IpAddr,
 }
 
 impl AuthenticatorClient {
     #[allow(clippy::too_many_arguments)]
-    fn new_type(
-        data_path: &Option<PathBuf>,
+    pub fn new(
         mixnet_listener: MixnetMessageBroadcastReceiver,
         mixnet_sender: MixnetMessageInputSender,
         our_nym_address: Recipient,
         auth_recipient: Recipient,
         auth_version: AuthenticatorVersion,
-        private_file_name: &str,
-        public_file_name: &str,
+        keypair: Arc<KeyPair>,
         ip_addr: IpAddr,
     ) -> Self {
-        let mut rng = OsRng;
-
-        let keypair = if let Some(data_path) = data_path {
-            let paths = KeyPairPath::new(
-                data_path.join(private_file_name),
-                data_path.join(public_file_name),
-            );
-            helpers::load_or_generate_keypair(&mut rng, paths)
-        } else {
-            KeyPair::new(&mut rng)
-        };
         Self {
             mixnet_listener,
             mixnet_sender,
@@ -79,50 +59,6 @@ impl AuthenticatorClient {
             keypair,
             ip_addr,
         }
-    }
-
-    pub fn new_entry(
-        data_path: &Option<PathBuf>,
-        mixnet_listener: MixnetMessageBroadcastReceiver,
-        mixnet_sender: MixnetMessageInputSender,
-        our_nym_address: Recipient,
-        auth_recipient: Recipient,
-        auth_version: AuthenticatorVersion,
-        ip_addr: IpAddr,
-    ) -> Self {
-        Self::new_type(
-            data_path,
-            mixnet_listener,
-            mixnet_sender,
-            our_nym_address,
-            auth_recipient,
-            auth_version,
-            DEFAULT_PRIVATE_ENTRY_WIREGUARD_KEY_FILENAME,
-            DEFAULT_PUBLIC_ENTRY_WIREGUARD_KEY_FILENAME,
-            ip_addr,
-        )
-    }
-
-    pub fn new_exit(
-        data_path: &Option<PathBuf>,
-        mixnet_listener: MixnetMessageBroadcastReceiver,
-        mixnet_sender: MixnetMessageInputSender,
-        our_nym_address: Recipient,
-        auth_recipient: Recipient,
-        auth_version: AuthenticatorVersion,
-        ip_addr: IpAddr,
-    ) -> Self {
-        Self::new_type(
-            data_path,
-            mixnet_listener,
-            mixnet_sender,
-            our_nym_address,
-            auth_recipient,
-            auth_version,
-            DEFAULT_PRIVATE_EXIT_WIREGUARD_KEY_FILENAME,
-            DEFAULT_PUBLIC_EXIT_WIREGUARD_KEY_FILENAME,
-            ip_addr,
-        )
     }
 
     pub async fn send_and_wait_for_response(
@@ -364,5 +300,92 @@ impl AuthenticatorClient {
         };
 
         Ok(gateway_data)
+    }
+
+    pub async fn query_bandwidth(&mut self) -> Result<Option<i64>> {
+        let query_message = match self.auth_version {
+            AuthenticatorVersion::V1 => return Err(Error::UnsupportedAuthenticatorVersion),
+            AuthenticatorVersion::V2 => ClientMessage::Query(Box::new(QueryMessageImpl {
+                pub_key: PeerPublicKey::new(self.keypair.public_key().to_bytes().into()),
+                version: AuthenticatorVersion::V2,
+            })),
+            AuthenticatorVersion::V3 => ClientMessage::Query(Box::new(QueryMessageImpl {
+                pub_key: PeerPublicKey::new(self.keypair.public_key().to_bytes().into()),
+                version: AuthenticatorVersion::V3,
+            })),
+            AuthenticatorVersion::V4 => ClientMessage::Query(Box::new(QueryMessageImpl {
+                pub_key: PeerPublicKey::new(self.keypair.public_key().to_bytes().into()),
+                version: AuthenticatorVersion::V4,
+            })),
+            AuthenticatorVersion::V5 => ClientMessage::Query(Box::new(QueryMessageImpl {
+                pub_key: PeerPublicKey::new(self.keypair.public_key().to_bytes().into()),
+                version: AuthenticatorVersion::V5,
+            })),
+            AuthenticatorVersion::UNKNOWN => return Err(Error::UnsupportedAuthenticatorVersion),
+        };
+        let response = self.send_and_wait_for_response(&query_message).await?;
+
+        let available_bandwidth = match response {
+            AuthenticatorResponse::RemainingBandwidth(remaining_bandwidth_response) => {
+                if let Some(available_bandwidth) =
+                    remaining_bandwidth_response.available_bandwidth()
+                {
+                    available_bandwidth
+                } else {
+                    return Ok(None);
+                }
+            }
+            _ => return Err(Error::InvalidGatewayAuthResponse),
+        };
+
+        let remaining_pretty = if available_bandwidth > 1024 * 1024 {
+            format!("{:.2} MB", available_bandwidth as f64 / 1024.0 / 1024.0)
+        } else {
+            format!("{} KB", available_bandwidth / 1024)
+        };
+        tracing::debug!(
+            "Remaining wireguard bandwidth with gateway {} for today: {}",
+            self.auth_recipient.gateway(),
+            remaining_pretty
+        );
+        if available_bandwidth < 1024 * 1024 {
+            tracing::warn!(
+                            "Remaining bandwidth is under 1 MB. The wireguard mode will get suspended after that until tomorrow, UTC time. The client might shutdown with timeout soon
+            "
+                        );
+        }
+        Ok(Some(available_bandwidth))
+    }
+
+    pub async fn top_up(&mut self, credential: CredentialSpendingData) -> Result<i64> {
+        let top_up_message = match self.auth_version {
+            AuthenticatorVersion::V3 => ClientMessage::TopUp(Box::new(v3::topup::TopUpMessage {
+                pub_key: PeerPublicKey::new(self.keypair.public_key().to_bytes().into()),
+                credential,
+            })),
+            // NOTE: looks like a bug here using v3. But we're leaving it as is since it's working
+            // and V4 is deprecated in favour of V5
+            AuthenticatorVersion::V4 => ClientMessage::TopUp(Box::new(v4::topup::TopUpMessage {
+                pub_key: PeerPublicKey::new(self.keypair.public_key().to_bytes().into()),
+                credential,
+            })),
+            AuthenticatorVersion::V5 => ClientMessage::TopUp(Box::new(v5::topup::TopUpMessage {
+                pub_key: PeerPublicKey::new(self.keypair.public_key().to_bytes().into()),
+                credential,
+            })),
+            AuthenticatorVersion::V1 | AuthenticatorVersion::V2 | AuthenticatorVersion::UNKNOWN => {
+                return Err(Error::UnsupportedAuthenticatorVersion);
+            }
+        };
+        let response = self.send_and_wait_for_response(&top_up_message).await?;
+
+        let remaining_bandwidth = match response {
+            AuthenticatorResponse::TopUpBandwidth(top_up_bandwidth_response) => {
+                top_up_bandwidth_response.available_bandwidth()
+            }
+            _ => return Err(Error::InvalidGatewayAuthResponse),
+        };
+
+        Ok(remaining_bandwidth)
     }
 }
