+++ nym-registration-client/src/config.rs
@@ -1,12 +1,10 @@
 // Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use nym_registration_common::NymNode;
-use std::path::PathBuf;
+use crate::builder::config::NymNodeWithKeys;
 
 pub struct RegistrationClientConfig {
-    pub(crate) entry: NymNode,
-    pub(crate) exit: NymNode,
+    pub(crate) entry: NymNodeWithKeys,
+    pub(crate) exit: NymNodeWithKeys,
     pub(crate) two_hops: bool,
-    pub(crate) data_path: Option<PathBuf>,
 }
