+++ sdk/rust/nym-sdk/src/mixnet.rs
@@ -63,7 +63,7 @@ pub use nym_credential_storage::{
     ephemeral_storage::EphemeralStorage as EphemeralCredentialStorage,
     models::StoredIssuedTicketbook, storage::Storage as CredentialStorage,
 };
-pub use nym_crypto::asymmetric::ed25519;
+pub use nym_crypto::asymmetric::{ed25519, x25519};
 pub use nym_network_defaults::NymNetworkDetails;
 pub use nym_socks5_client_core::config::Socks5;
 pub use nym_sphinx::{
